application/repositories/video_repository.go
package repositories

import (
	"video-encoder/domain"
	"github.com/jinzhu/gorm"
	uuid "github.com/satori/go.uuid"
	"fmt"
)

type VideoRepository interface {
	Insert(video *domain.Video) (*domain.Video, error)
	Find(id string) (*domain.Video, error)
}

type VideoRepositoryDb struct {
	Db *gorm.DB
}

func NewVideoRepository(db *gorm.DB) *VideoRepositoryDb {
	return &VideoRepositoryDb{Db:db}
}

func (repository VideoRepositoryDb) Insert(video *domain.Video) (*domain.Video, error) {
	if video.ID == "" {
		video.ID = uuid.NewV4().String()
	}

	err := repository.Db.Create(video).Error

	if err != nil {
		return nil, err
	}

	return video, nil
}

func (repository VideoRepositoryDb) Find(id string) (*domain.Video, error) {
	var video domain.Video
	repository.Db.Preload("Jobs").First(&video, "id = ?", id)

	if video.ID == "" {
		return nil, fmt.Errorf("video does not exists")
	}

	return &video, nil
}